# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AllMart API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/messages:
    get:
      operationId: listChatMessages
      tags: [chat]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatTurnResponse"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        price: { type: number }
        originalPrice:
          type: number
          nullable: true
        currency: { type: string }
        imageUrl: { type: string }
        images:
          type: array
          items: { type: string }
        colors:
          type: array
          items: { type: string }
        productType: { type: string }
        rating: { type: number }
        stock: { type: integer }
        sellerName: { type: string }
        tags:
          type: array
          items: { type: string }
      required: [id, name, description, category, price, currency, imageUrl, images, colors, productType, rating, stock, sellerName, tags]
...
    CartItem:
      type: object
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        product: { $ref: "#/components/schemas/Product" }
      required: [productId, quantity, product]
...
    OrderItem:
      type: object
      properties:
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        imageUrl: { type: string }
      required: [productId, productName, quantity, unitPrice, imageUrl]
    Order:
      type: object
      properties:
        id: { type: integer }
        status: { type: string, enum: [placed, confirmed, dispatched, delivered, cancelled] }
        total: { type: number }
        currency: { type: string }
        trackingCode: { type: string }
        shippingAddress: { type: string }
        receiverName:
          type: string
          nullable: true
        receiverEmail:
          type: string
          nullable: true
        receiverPhone:
          type: string
          nullable: true
        cashbackCode:
          type: string
          nullable: true
        cashbackDiscount:
          type: number
          nullable: true
        placedBy: { type: string, enum: [user, ai] }
        createdAt: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
      required: [id, status, total, currency, trackingCode, shippingAddress, placedBy, createdAt, items]
...
    ChatMessage:
      type: object
      properties:
        id: { type: integer }
        role: { type: string, enum: [user, assistant, system] }
        content: { type: string }
        createdAt: { type: string }
        productSuggestions:
          type: array
          items: { $ref: "#/components/schemas/Product" }
      required: [id, role, content, createdAt]
    SendChatMessageRequest:
      type: object
      properties:
        content: { type: string, minLength: 1 }
        confirmOrder:
          type: boolean
          description: When true, AI is authorized to place order with current cart
        shippingAddress: { type: string }
        paymentMethod:
          type: string
          enum: [stripe, bank_transfer, pay_on_delivery]
          description: Payment method chosen by the shopper
      required: [content]
    ChatBankDetails:
      type: object
      properties:
        bankName: { type: string }
        accountName: { type: string }
        accountNumber: { type: string }
        routingNumber: { type: string }
        email: { type: string }
      required: [bankName, accountName, accountNumber]
    ChatTurnResponse:
      type: object
      properties:
        userMessage: { $ref: "#/components/schemas/ChatMessage" }
        assistantMessage: { $ref: "#/components/schemas/ChatMessage" }
        suggestedProducts:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        addedToCart:
          type: array
          items: { $ref: "#/components/schemas/CartItem" }
        placedOrder:
          $ref: "#/components/schemas/Order"
        needsConfirmation:
          type: boolean
        needsShippingAddress:
          type: boolean
        stripeCheckoutUrl:
          type: string
          description: Stripe checkout URL when paymentMethod=stripe
        bankDetails:
          $ref: "#/components/schemas/ChatBankDetails"
          description: Bank details when paymentMethod=bank_transfer
      required: [userMessage, assistantMessage, suggestedProducts, addedToCart, needsConfirmation, needsShippingAddress]